fix keywords and wordchars
Currently filetypes.r has a rather random collection of keywords for R files. 

This commit fixes the primary keywords (as used in RStudio, an IDE specialized for R), and also adds the appropriate list of built-in constants. See https://github.com/rstudio/rstudio/blob/master/src/gwt/acesupport/acemode/r_highlight_rules.js for the list of keywords used in RStudio. It also comments out `package_other` line, as this list serves no useful purpose. 

Furthermore, the commit adds to `wordchars` the `.` (dot). In R the set of symbols which can be used in object names is `[A-Za-z0-9]` and `_` and `.` (with code portability and platform compatibility in mind). For more gruesome details see: http://cran.r-project.org/doc/manuals/r-release/R-intro.html#R-commands_003b-case-sensitivity-etc . RStudio also treats the dot as part of the "word". 

For some background see: http://landroni.wordpress.com/2013/07/27/using-geany-for-programming-in-r/