Added Protocol Buffers (protobuf) filetypes
Google's Protocol Buffers, also known as protobuf, is a tool for aiding
in the process of serializing structured data. Protobuf has employed its
own little language which allows it to be compiled down into almost any
language. With the protobuf language getting advanced at times, it makes
having file handling and syntax highlighting in editor very valuable.

This implements the built-in protobuf filetypes and highlight mappings.