i18n: don't restrict LINGUAS at configure time
Don't use the $LINGUAS at configure time to set ALL_LINGUAS, and let
the translations Makefile properly honor the $LINGUAS itself, which it
already does better than we do, cleanly ignoring translations it
doesn't know.

If we do set ALL_LINGUAS=$LINGUAS, it will result in a build failure if
we do not have a translation for some of the language(s) specified in
$LINGUAS, and would make it impossible to build other languages without
re-configuring.  So, just drop that part and let the build-time support
kick in.

Closes #507.