Fixes for tag processing of COBOL files
  Updated the current tag parser of COBOL files to better handle the
various keywords and to suppress false positives of various types.

NOTE: code folding for COBOL files only works if the sequence number
area is dropped , the indicator area begins in column 1 and those
statements designated for area A begin in column 2. As such, the tag
parsing is set up to deal with that type of file structure. The test for
COBOL tagging has been changed to reflect this.