Expose the tm_workspace_update() method to plugins
This makes it possible to call tm_source_file_update() in loop for multiple
files without workspace update for each of them. At the end of the loop,
tm_workspace_update() can be called just once. This makes a huge performance
impact when the number of files is large.

Also when recursing, set force parameter to TRUE for the child (file) objects
- setting to FALSE doesn't make sense since children are always files
and for the FALSE argument the update isn't performed.