autotools: Only uninstall what we actually installed
Only manually uninstall the files we manually installed, as
uninstall-local might be run in parallel to Automake's own uninstall
targets.

We don't uninstall the directories as some docs seems to suggest
`rmdir` isn't really portable.  Anyway, Automake don't uninstall
directories either so it's not a real problem.

An alternative solution would be to keep the recursive removal of
`$(DOCDIR)` but move it to `uninstall-hook`, which is guaranteed to be
run after other uninstall rules.  However, recursive deletion is not
always sensible as it might remove files we didn't install, e.g. if the
user added them manually.