"Execute commands" -> "Comandos de ejecución"
"Execute commands" (from the Set Build Commands dialog) is translated to Spanish as if it were a sentence (like "Execute these commands"), but it's actually a name ("<language> commands", "Independent commands", and "Execute commands"; so it means "Commands for 'execute'").  Therefore, +"Comandos de ejecución" is a better translation than -"Ejecutar comandos".