TM: Consider vartype when sorting and de-duplicating global tags
Don't consider two global tags as identical if their vartype is
different.

This fixes tests/ctags/keyword_event.cs on some systems (seen on a
Debian 6 VM), probably because of a sort instability of some sort.
The problem with this test case is that it has two identical tags but
for their return type, and the picked one was different.

Note that the results for test case tests/ctags/bug507864.c don't seem
really correct, but that's an unrelated issue.