The GEANY_*_SYMBOL flags also need to be defined for C++
In Waf, we need to add the new flags to CFLAGS *and* CXXFLAGS otherwise
the Scintilla build would break.
To keep it more reabable, we first populate a temporary list
geany_symbol_flags which is then added to CFLAGS and CXXFLAGS.

Concerning the quote FIXME: no further qoting is necessary here.