Use lsof to get cwd of shell when /proc/pid/cwd isn't available
lsof can be used to get cwd of a process and it's available
on basically every UNIX. In addition, it offers the "output for
other programs" mode which is kind of standardized and
shouldn't change in the future. I have tested the lsof command
on linux, PC BSD and OS X and both the parameters and
output are mostly identical everywhere (on PC BSD there's
an additional 'f' field). The patch performs some checks if
the output corresponds to what we expect.

In addition, it tries to find procfs on alternative paths under
various BSD systems (code taken from xfce terminal).