Add a "geany-document" tag to Scintilla widgets
This is not strictly needed but it makes the program more robust
against plugins breaking the UI by re-arranging the notebook pages
behind our back. This is specifically useful for the Overview plugin
which puts another Scintilla widget to the left of the normal editor
one, and the previous search by scanning the notebook page widgets
found it instead of its own, and trouble ensued.

Since the tag needs to contain something, set it to the related
GeanyDocument so we can make use of it from `document_find_by_sci`.