Alter build.c to use the new spawning module
The wrong and misleading "Failed to change the working directory" is
changed to a generic "Invalid working directory".

Under Windows, the build commands are run directly as command lines,
while under Unix, /bin/sh is used - for compatibility, and because
that's probably what a *nix user expects.

The run script is always run as a command line, since it contains it's
own shell, and because that lets us specify %c under Windows. In fact,
%c (or "%c") should probably be the default, since CreateProcess()
runs batch files with the default shell, instead of some our value.

Also changed build_spawn_cmd() and build_run_cmd() to void, since
their return values are not used anywhere.