Remove the possibility to specify base directory and simplify project creation
The possibility to have the base path at a different place than
the project file makes things more complicated, confusing for users
and impossible to specify open files relatively to the project
file.

For these reasons, this patch removes the possibility to specify the
base directory - after this patch the base directory is always the
directory where the project file is stored.

Since the base path is absolute now, remove all the code dealing
with the relative path.

Because the project file is now always in the root directory of the
project, we can simplify project opening and creation - it can become
just a simple directory opening. If the project file is in the
directory, open the project; otherwise create a new project. To find
the project, the project file name has to be always the same -
use "project.geany" as the project file name.

To better distinguish project creation from project opening and
to simplify initial project configuration, open the project properties
dialog when a new project is created.

Since for a user project is basically a directory now (the user doesn't
have to deal with the project.geany file which is created automatically),
show only directories in recent projects list. Use new "recent_project_dirs"
config file variable so old project files in the list aren't mixed with
the new project dirs.

Finally, this patch adds a new dialogs_show_open_dialog() function
to open file/directory dialogs which is then reused by other parts
of the code and eliminates code duplication.

The rest of the patch is a lot of (now unused) code removal.