icon: Adjust the color of the inside of the bottleneck
The (almost) black color of the inside of the bottleneck is a bit too high
contrast and doesn't fit the rest of the icon (the only black part of it).
Use the "lamp yellow" as the base and lower the L in HSL representation of
the color so it just looks as darker lamp material.