Start bash as a login shell in VTE under OS X
Graphical applications on OS X are started using launchd and don't inherit
shell's environment variables so we don't get the PS1 value. Do the same
as the OS X terminal emulator which runs the shell as a login shell (and yes,
on OS X .bash_profile is read for shells started in the terminal application
instead of .bashrc).