Disable empty tree view focus in message windows
When the tree view inside Status/Compiler/Messages is empty and the tab is
clicked, the widget which is inside the notebook (in our case tree view)
gets focused. If there are some rows in it, a row gets focused. If not,
however, the whole tree view gets focused. Because some GTK themes show
focused widget as a colored rectangle, the whole tree view is painted in
the given color.

Disable focus for the tree view if it's empty and reenable it when something
is inserted.