Tell users how to clear the terminal in the error message
This is not completely obvious - when I first saw the message, I started
pressing backspace which really doesn't help.

Also clarify and shorten the message a bit - in "Could not execute the file
in the VTE because it probably contains a command" it's not clear if it's
the file or VTE which contains the command. Also use "terminal" instead
of "VTE" which is more user-friendly.