Fix language check in tm_workspace_find()
The tags_lang variable is set from the first tag in the found array but
the array may actually contain tags from several languages. This may
lead to two things:

1. Goto tag definition goes to a tag from a different filetype
2. Worse, the first tag is from a different language than the current file
and we get a message that no tag was found

Get lang for every tag in the array and rename tags_lang to tag_lang.